Refactoring to take in account new module ops_ste
