+++ brevitas/core/bit_width.py
@@ -47,7 +47,8 @@ from torch.nn import Parameter
 
 import brevitas.config as config
 from brevitas.utils.python_utils import AutoName
-from brevitas.function.ops import tensor_clamp_ste, tensor_clamp
+from brevitas.function.ops import tensor_clamp
+from brevitas.function.ops_ste import tensor_clamp_ste
 from .restrict_val import RestrictValueOpImplType, RestrictValueType, RestrictValue, FloatToIntImplType
