+++ brevitas/nn/hadamard_classifier.py
@@ -8,7 +8,8 @@ import torch
 from scipy.linalg import hadamard
 
 
-from brevitas.function.ops import ceil_ste, max_uint
+from brevitas.function.ops import ceil_ste
+from brevitas.function.ops_ste import max_uint
 from .quant_layer import QuantLayer
